reorganize css
