+++ frontend/src/pages/Dashboard/HomeMenu.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useRef, useEffect } from "react";
 import HomeDecklist from "./HomeDecklist";
 import HomeMenuHeader from "./HomeMenuHeader";
-import styles from "./Dashboard.module.css";
+import styles from "./styles/Dashboard.module.css";
 
 const HomeMenu: React.FC = () => {
   const [showInput, setShowInput] = useState(false);
