+++ frontend/src/pages/StudySession/FlashcardControls.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./FlashcardControls.module.css";
+import styles from "./styles/FlashcardControls.module.css";
 
 interface FlashcardControlsProps {
   prevCard: () => void;
